Refactored world and shape functions
