+++ src/3d/utils.ts
@@ -1,0 +1,6 @@
+  // Convert degrees to radians
+  const convertToRadians = (degrees: number) => {
+    return degrees * (Math.PI / 180);
+  };
+
+  export { convertToRadians };
